REFINE: remove unused `isUserInRole` utility and clean up unused imports, replace Dutch naming with English counterparts, and consolidate redundant product and auction forms.
