+++ LeafBidWeb/src/app/auction/[id]/page.tsx
