+++ LeafBidWeb/src/components/header/header.tsx
@@ -8,6 +8,8 @@ import ThemeInitializer, {getTheme, toggleTheme} from "./theme";
 import {useEffect, useState} from "react";
 import {LoggedInResponse} from "@/types/User/Auth/LoggedInResponse";
 import {useRouter} from "next/navigation";
+import {isUserInRole} from "@/utils/isUserInRole";
+import {parseRole, Roles} from "@/enums/Roles";
 
 interface HeaderProps {
     returnOption?: boolean;
@@ -16,17 +18,9 @@ interface HeaderProps {
 export default function Header({returnOption = false}: HeaderProps) {
     const router = useRouter();
     const [theme, setTheme] = useState<"dark" | "light">("light");
-    const [isAuctioneer, setIsAuctioneer] = useState(false);
-    const [isProvider, setIsProvider] = useState(false);
-    const [isAdmin, setIsAdmin] = useState(false);
 
     useEffect(() => {
         setTheme(getTheme());
-    }, []);
-
-    // call fetchUserData on mount (top-level useEffect)
-    useEffect(() => {
-        fetchUserData();
     }, []);
 
     const onToggleTheme = () => {
@@ -44,7 +38,9 @@ export default function Header({returnOption = false}: HeaderProps) {
                 headers: {"Content-Type": "application/json"},
             });
 
-            if (!res.ok) throw new Error("Logout mislukt");
+            if (!res.ok) {
+                throw new Error("Logout mislukt, error: " + await res.json() + "");
+            }
 
             document.cookie = "session=; Max-Age=0; path=/";
             document.cookie = "PHPSESSID=; Max-Age=0; path=/";
@@ -59,45 +55,6 @@ export default function Header({returnOption = false}: HeaderProps) {
         }
     };
 
-    interface UserData {
-        roles?: (string | null)[] | null;
-        role?: string | null;
-        isAuctioneer?: boolean | null;
-        isProvider?: boolean | null;
-        isAdmin?: boolean | null;
-    }
-
-
-    const determineRoles = (userData?: UserData | null) => {
-        if (!userData) return {auctioneer: false, provider: false, admin: false};
-
-        const lower = (v?: string | null) => (typeof v === "string" ? v.toLowerCase() : "");
-        const rolesArray = Array.isArray(userData.roles)
-            ? userData.roles
-                .filter((r): r is string => typeof r === "string")
-                .map((r) => r.toLowerCase())
-            : [];
-
-        const auctioneer =
-            userData.isAuctioneer === true ||
-            lower(userData.role) === "auctioneer" ||
-            rolesArray.includes("auctioneer");
-
-        const provider =
-            userData.isProvider === true ||
-            lower(userData.role) === "provider" ||
-            rolesArray.includes("provider") ||
-            rolesArray.includes("aanvoerder");
-
-        const admin =
-            userData.isAdmin === true ||
-            lower(userData.role) === "admin" ||
-            rolesArray.includes("admin");
-
-        return {auctioneer, provider, admin};
-    };
-
-
     const fetchUserData = () => {
         // Send request to localhost:5001/api/v2/User/me
         fetch("http://localhost:5001/api/v2/User/me", {
@@ -110,17 +67,10 @@ export default function Header({returnOption = false}: HeaderProps) {
                     // Set localStorage
                     localStorage.setItem("loggedIn", "true");
                     localStorage.setItem("userData", JSON.stringify(response.userData));
-                    const {auctioneer, provider, admin} = determineRoles(response.userData);
-                    setIsAuctioneer(auctioneer);
-                    setIsProvider(provider);
-                    setIsAdmin(admin);
                 } else {
                     // Remove localStorage
                     localStorage.setItem("loggedIn", "false");
                     localStorage.removeItem("userData");
-                    setIsAuctioneer(false);
-                    setIsProvider(false);
-                    setIsAdmin(false);
                 }
             })
             .finally(() => {
@@ -131,32 +81,18 @@ export default function Header({returnOption = false}: HeaderProps) {
             });
     };
 
-    const checkLoggedInState = () => {
+    const checkLoggedInState = async () => {
         if (
             localStorage.getItem("loggedIn") &&
             localStorage.getItem("loggedIn") === "false"
         ) {
-            logout();
-        } else {
-            const stored = localStorage.getItem("userData");
-            if (stored) {
-                try {
-                    const parsed = JSON.parse(stored);
-                    const {auctioneer, provider, admin} = determineRoles(parsed);
-                    setIsAuctioneer(auctioneer);
-                    setIsProvider(provider);
-                    setIsAdmin(admin);
-                } catch {
-                    setIsAuctioneer(false);
-                    setIsProvider(false);
-                    setIsAdmin(false);
-                }
-            }
+            await logout();
         }
     };
+
     useEffect(() => {
         fetchUserData();
-    }, []);
+    });
     return (
         <header>
             <ThemeInitializer/>
@@ -181,18 +117,21 @@ export default function Header({returnOption = false}: HeaderProps) {
                     </Link>
                 )}
                 <div className={s.clickables}>
-                    {(isProvider || isAdmin) && (
-                        <Link href="/product/toevoegen" className={s.link}>
-                            Product toevoegen
+                    {(isUserInRole(parseRole(Roles.Provider))) && (
+                        <Link href="/product/register" className={s.link}>
+                            Product registreren
                         </Link>
                     )}
 
-                    {(isAuctioneer || isAdmin) && (
+                    {(isUserInRole(parseRole(Roles.Auctioneer))) && (
                         <>
-                            <Link href="/veiling/toevoegen" className={s.link}>
+                            <Link href="/product/add" className={s.link}>
+                                Product aanmaken
+                            </Link>
+                            <Link href="/auction/add" className={s.link}>
                                 Veiling toevoegen
                             </Link>
-                            <Link href="/veiling/veilingmeesterOverzicht" className={s.link}>
+                            <Link href="/auction/dashboard" className={s.link}>
                                 Veilingmeester overzicht
                             </Link>
                         </>
