+++ LeafBidWeb/src/types/Auction/Location.ts
@@ -1,4 +1,4 @@
-export interface Locatie {
+export interface Location {
     locatieId: number;
     locatieNaam: string;
 }
